+++ .github/workflows/test.lua
@@ -1,63 +1,19 @@
 hook.Add("Tick", "Template_CloseServer", engine.CloseServer)
 require("dotnet")
 
-function donothingwithit(something) end
-
-local function benchmark(name, fn)
-	local startTime = SysTime()
-	fn()
-	local elapsed = SysTime()-startTime
-	print(name.." "..elapsed)
-	return elapsed
-end
-
-local function benchmark_thousand(name, fn)
-	local startTime = SysTime()
-	for i=0,1000 do
-		fn()
-	end
-	local elapsed = SysTime()-startTime
-	print(name.."_thousand"..elapsed)
-	return elapsed
-end
-
-local function math_Approach()
-	benchmark("math_Approach", function()
-		local current = 0
-		for i=0,10000 do
-			current = math.Approach(current,100,0.01)
-		end
-		print(current)
-	end)
-end
-
-local function math_Distance()
-	benchmark("math_Distance", function()
-		local current = 0
-		for i=0,10000 do
-			current = math.Distance(current,0,200,100)
-		end
-		print(current)
-	end)
-end
-
 local function run_test()
-	local module_loaded = dotnet.load("NativeMath")
+	local module_loaded = dotnet.load("HashCalc")
 	assert(module_loaded)
 
 	-----------------------
 
-	math_Approach()
-	math_Distance()
-
-	NativeMath()
-
-	math_Approach()
-	math_Distance()
+	local sha256 = HashCalc.CalculateSHA256("test")
+	print("SHA256 = " .. sha256)
+	assert(sha256=="9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08")
 
 	-----------------------
 
-	local module_unloaded = dotnet.unload("NativeMath")
+	local module_unloaded = dotnet.unload("HashCalc")
 	assert(module_unloaded)
 end
