+++ HashCalc/HashCalc.csproj
