+++ HashCalc/nuget.config
